requirements.txt
tensorflow-cpu
mlflow
dvc
fastapi
uvicorn
opencv-python
numpy
pandas
python-multipart
